move KeyLockedLazyCache to its own file
